+++ packages/shared_preferences/shared_preferences_android/android/src/test/kotlin/io/flutter/plugins/sharedpreferences/SharedPreferencesTest.kt
@@ -11,11 +11,12 @@ import androidx.preference.PreferenceManager
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import io.flutter.embedding.engine.plugins.FlutterPlugin
-import io.flutter.plugin.common.BinaryMessenger
 import io.mockk.every
 import io.mockk.mockk
 import java.io.ByteArrayOutputStream
 import java.io.ObjectOutputStream
+import kotlinx.coroutines.async
+import kotlinx.coroutines.runBlocking
 import org.junit.Assert
 import org.junit.Assert.assertThrows
 import org.junit.Test
@@ -49,14 +50,10 @@ internal class SharedPreferencesTest {
 
   private fun pluginSetup(options: SharedPreferencesPigeonOptions): SharedPreferencesAsyncApi {
     val plugin = SharedPreferencesPlugin()
-    val binaryMessenger = mockk<BinaryMessenger>()
     val flutterPluginBinding = mockk<FlutterPlugin.FlutterPluginBinding>()
-    every { flutterPluginBinding.binaryMessenger } returns binaryMessenger
     every { flutterPluginBinding.applicationContext } returns testContext
     plugin.onAttachedToEngine(flutterPluginBinding)
-    val backend =
-        SharedPreferencesBackend(
-            flutterPluginBinding.binaryMessenger, flutterPluginBinding.applicationContext)
+    val backend = SharedPreferencesBackend(flutterPluginBinding.applicationContext)
     return if (options.useDataStore) {
       plugin
     } else {
@@ -65,35 +62,35 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testSetAndGetBoolWithDataStore() {
+  fun testSetAndGetBoolWithDataStore() = runBlocking {
     val plugin = pluginSetup(dataStoreOptions)
     plugin.setBool(boolKey, testBool, dataStoreOptions)
     Assert.assertEquals(plugin.getBool(boolKey, dataStoreOptions), testBool)
   }
 
   @Test
-  fun testSetAndGetStringWithDataStore() {
+  fun testSetAndGetStringWithDataStore() = runBlocking {
     val plugin = pluginSetup(dataStoreOptions)
     plugin.setString(stringKey, testString, dataStoreOptions)
     Assert.assertEquals(plugin.getString(stringKey, dataStoreOptions), testString)
   }
 
   @Test
-  fun testSetAndGetIntWithDataStore() {
+  fun testSetAndGetIntWithDataStore() = runBlocking {
     val plugin = pluginSetup(dataStoreOptions)
     plugin.setInt(intKey, testInt, dataStoreOptions)
     Assert.assertEquals(plugin.getInt(intKey, dataStoreOptions), testInt)
   }
 
   @Test
-  fun testSetAndGetDoubleWithDataStore() {
+  fun testSetAndGetDoubleWithDataStore() = runBlocking {
     val plugin = pluginSetup(dataStoreOptions)
     plugin.setDouble(doubleKey, testDouble, dataStoreOptions)
     Assert.assertEquals(plugin.getDouble(doubleKey, dataStoreOptions), testDouble)
   }
 
   @Test
-  fun testSetAndGetStringListWithDataStore() {
+  fun testSetAndGetStringListWithDataStore() = runBlocking {
     val plugin = pluginSetup(dataStoreOptions)
     plugin.setEncodedStringList(listKey, testList, dataStoreOptions)
     val result = plugin.getStringList(listKey, dataStoreOptions)
@@ -102,7 +99,7 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testSetAndGetStringListWithDataStoreRedirectsForPlatformEncoded() {
+  fun testSetAndGetStringListWithDataStoreRedirectsForPlatformEncoded() = runBlocking {
     val plugin = pluginSetup(dataStoreOptions)
     plugin.setDeprecatedStringList(listKey, listOf(""), dataStoreOptions)
     val result = plugin.getStringList(listKey, dataStoreOptions)
@@ -111,7 +108,7 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testSetAndGetStringListWithDataStoreReportsRawString() {
+  fun testSetAndGetStringListWithDataStoreReportsRawString() = runBlocking {
     val plugin = pluginSetup(dataStoreOptions)
     plugin.setString(listKey, testString, dataStoreOptions)
     val result = plugin.getStringList(listKey, dataStoreOptions)
@@ -120,7 +117,7 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testGetKeysWithDataStore() {
+  fun testGetKeysWithDataStore() = runBlocking {
     val plugin = pluginSetup(dataStoreOptions)
     plugin.setBool(boolKey, testBool, dataStoreOptions)
     plugin.setString(stringKey, testString, dataStoreOptions)
@@ -134,7 +131,7 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testClearWithDataStore() {
+  fun testClearWithDataStore() = runBlocking {
     val plugin = pluginSetup(dataStoreOptions)
     plugin.setBool(boolKey, testBool, dataStoreOptions)
     plugin.setString(stringKey, testString, dataStoreOptions)
@@ -152,7 +149,7 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testGetAllWithDataStore() {
+  fun testGetAllWithDataStore() = runBlocking {
     val plugin = pluginSetup(dataStoreOptions)
     plugin.setBool(boolKey, testBool, dataStoreOptions)
     plugin.setString(stringKey, testString, dataStoreOptions)
@@ -170,7 +167,7 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testClearWithAllowListWithDataStore() {
+  fun testClearWithAllowListWithDataStore() = runBlocking {
     val plugin = pluginSetup(dataStoreOptions)
     plugin.setBool(boolKey, testBool, dataStoreOptions)
     plugin.setString(stringKey, testString, dataStoreOptions)
@@ -188,7 +185,7 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testGetAllWithAllowListWithDataStore() {
+  fun testGetAllWithAllowListWithDataStore() = runBlocking {
     val plugin = pluginSetup(dataStoreOptions)
     plugin.setBool(boolKey, testBool, dataStoreOptions)
     plugin.setString(stringKey, testString, dataStoreOptions)
@@ -206,35 +203,35 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testSetAndGetBoolWithSharedPreferences() {
+  fun testSetAndGetBoolWithSharedPreferences() = runBlocking {
     val plugin = pluginSetup(sharedPreferencesOptions)
     plugin.setBool(boolKey, testBool, sharedPreferencesOptions)
     Assert.assertEquals(plugin.getBool(boolKey, sharedPreferencesOptions), testBool)
   }
 
   @Test
-  fun testSetAndGetStringWithSharedPreferences() {
+  fun testSetAndGetStringWithSharedPreferences() = runBlocking {
     val plugin = pluginSetup(sharedPreferencesOptions)
     plugin.setString(stringKey, testString, sharedPreferencesOptions)
     Assert.assertEquals(plugin.getString(stringKey, sharedPreferencesOptions), testString)
   }
 
   @Test
-  fun testSetAndGetIntWithSharedPreferences() {
+  fun testSetAndGetIntWithSharedPreferences() = runBlocking {
     val plugin = pluginSetup(sharedPreferencesOptions)
     plugin.setInt(intKey, testInt, sharedPreferencesOptions)
     Assert.assertEquals(plugin.getInt(intKey, sharedPreferencesOptions), testInt)
   }
 
   @Test
-  fun testSetAndGetDoubleWithSharedPreferences() {
+  fun testSetAndGetDoubleWithSharedPreferences() = runBlocking {
     val plugin = pluginSetup(sharedPreferencesOptions)
     plugin.setDouble(doubleKey, testDouble, sharedPreferencesOptions)
     Assert.assertEquals(plugin.getDouble(doubleKey, sharedPreferencesOptions), testDouble)
   }
 
   @Test
-  fun testSetAndGetStringListWithSharedPreferences() {
+  fun testSetAndGetStringListWithSharedPreferences() = runBlocking {
     val plugin = pluginSetup(sharedPreferencesOptions)
     plugin.setEncodedStringList(listKey, testList, sharedPreferencesOptions)
     val result = plugin.getStringList(listKey, sharedPreferencesOptions)
@@ -243,7 +240,7 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testSetAndGetStringListWithSharedPreferencesRedirectsForPlatformEncoded() {
+  fun testSetAndGetStringListWithSharedPreferencesRedirectsForPlatformEncoded() = runBlocking {
     val plugin = pluginSetup(sharedPreferencesOptions)
     plugin.setDeprecatedStringList(listKey, listOf(""), sharedPreferencesOptions)
     val result = plugin.getStringList(listKey, sharedPreferencesOptions)
@@ -252,7 +249,7 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testSetAndGetStringListWithSharedPreferencesReportsRawString() {
+  fun testSetAndGetStringListWithSharedPreferencesReportsRawString() = runBlocking {
     val plugin = pluginSetup(sharedPreferencesOptions)
     plugin.setString(listKey, testString, sharedPreferencesOptions)
     val result = plugin.getStringList(listKey, sharedPreferencesOptions)
@@ -261,7 +258,7 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testGetKeysWithSharedPreferences() {
+  fun testGetKeysWithSharedPreferences() = runBlocking {
     val plugin = pluginSetup(sharedPreferencesOptions)
     plugin.setBool(boolKey, testBool, sharedPreferencesOptions)
     plugin.setString(stringKey, testString, sharedPreferencesOptions)
@@ -275,7 +272,7 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testClearWithSharedPreferences() {
+  fun testClearWithSharedPreferences() = runBlocking {
     val plugin = pluginSetup(sharedPreferencesOptions)
     plugin.setBool(boolKey, testBool, sharedPreferencesOptions)
     plugin.setString(stringKey, testString, sharedPreferencesOptions)
@@ -293,7 +290,7 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testGetAllWithSharedPreferences() {
+  fun testGetAllWithSharedPreferences() = runBlocking {
     val plugin = pluginSetup(sharedPreferencesOptions)
     plugin.setBool(boolKey, testBool, sharedPreferencesOptions)
     plugin.setString(stringKey, testString, sharedPreferencesOptions)
@@ -311,7 +308,7 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testClearWithAllowListWithSharedPreferences() {
+  fun testClearWithAllowListWithSharedPreferences() = runBlocking {
     val plugin = pluginSetup(sharedPreferencesOptions)
     plugin.setBool(boolKey, testBool, sharedPreferencesOptions)
     plugin.setString(stringKey, testString, sharedPreferencesOptions)
@@ -329,12 +326,12 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testGetAllWithAllowListWithSharedPreferences() {
+  fun testGetAllWithAllowListWithSharedPreferences() = runBlocking {
     val plugin = pluginSetup(sharedPreferencesOptions)
     plugin.setBool(boolKey, testBool, sharedPreferencesOptions)
     plugin.setString(stringKey, testString, sharedPreferencesOptions)
     plugin.setInt(intKey, testInt, sharedPreferencesOptions)
-    plugin.setDouble(doubleKey, testDouble, sharedPreferencesOptions)
+    plugin.setDouble(doubleKey, testDouble, dataStoreOptions)
     plugin.setEncodedStringList(listKey, testList, sharedPreferencesOptions)
 
     val all = plugin.getAll(listOf(boolKey, stringKey), sharedPreferencesOptions)
@@ -347,7 +344,7 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testSharedPreferencesWithMultipleFiles() {
+  fun testSharedPreferencesWithMultipleFiles() = runBlocking {
     val plugin = pluginSetup(sharedPreferencesOptions)
     val optionsWithNewFile =
         SharedPreferencesPigeonOptions(useDataStore = false, fileName = "test_file")
@@ -358,7 +355,7 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testSharedPreferencesDefaultFile() {
+  fun testSharedPreferencesDefaultFile() = runBlocking {
     val defaultPreferences: SharedPreferences =
         PreferenceManager.getDefaultSharedPreferences(testContext)
     defaultPreferences.edit().putString(stringKey, testString).commit()
@@ -367,7 +364,7 @@ internal class SharedPreferencesTest {
   }
 
   @Test
-  fun testUnexpectedClassDecodeThrows() {
+  fun testUnexpectedClassDecodeThrows(): Unit = runBlocking {
     // Only String should be allowed in an encoded list.
     val badList = listOf(1, 2, 3)
     // Replicate the behavior of ListEncoder.encode, but with a non-List<String> list.
@@ -382,7 +379,62 @@ internal class SharedPreferencesTest {
     // Inject the bad pref as a string, as that is how string lists are stored internally.
     plugin.setString(badListKey, badPref, dataStoreOptions)
     assertThrows(ClassNotFoundException::class.java) {
-      plugin.getPlatformEncodedStringList(badListKey, dataStoreOptions)
+      runBlocking { plugin.getPlatformEncodedStringList(badListKey, dataStoreOptions) }
+    }
+  }
+
+  @Test
+  fun testExecutesOnBackgroundThreadSeriallyInFifoOrder() = runBlocking {
+    for (options in listOf(dataStoreOptions, sharedPreferencesOptions)) {
+      val callerThread = Thread.currentThread()
+      val executionThreads = java.util.Collections.synchronizedList(mutableListOf<Thread>())
+      val executionOrder = java.util.Collections.synchronizedList(mutableListOf<Int>())
+      val activeCount = java.util.concurrent.atomic.AtomicInteger(0)
+      val maxConcurrent = java.util.concurrent.atomic.AtomicInteger(0)
+
+      val trackingEncoder =
+          object : SharedPreferencesListEncoder {
+            override fun encode(list: List<String>): String {
+              val currentActive = activeCount.incrementAndGet()
+              maxConcurrent.updateAndGet { maxOf(it, currentActive) }
+              executionThreads.add(Thread.currentThread())
+              val index = list.first().toInt()
+              if (index == 0) {
+                Thread.sleep(50)
+              }
+              executionOrder.add(index)
+              activeCount.decrementAndGet()
+              return ListEncoder().encode(list)
+            }
+
+            override fun decode(listString: String): List<String> = ListEncoder().decode(listString)
+          }
+
+      val plugin = SharedPreferencesPlugin(trackingEncoder)
+      val flutterPluginBinding = mockk<FlutterPlugin.FlutterPluginBinding>()
+      every { flutterPluginBinding.applicationContext } returns testContext
+      plugin.onAttachedToEngine(flutterPluginBinding)
+
+      val api: SharedPreferencesAsyncApi =
+          if (options.useDataStore) {
+            plugin
+          } else {
+            SharedPreferencesBackend(testContext, trackingEncoder)
+          }
+
+      val jobs =
+          (0 until 5).map { i ->
+            async {
+              @Suppress("DEPRECATION")
+              api.setDeprecatedStringList("key_$i", arrayListOf("$i", "val"), options)
+            }
+          }
+      jobs.forEach { it.await() }
+      api.clear(null, options)
+
+      Assert.assertEquals(listOf(0, 1, 2, 3, 4), executionOrder)
+      Assert.assertEquals(1, maxConcurrent.get())
+      Assert.assertTrue(executionThreads.all { it != callerThread })
     }
   }
 }