+++ packages/shared_preferences/shared_preferences_android/pigeons/messages_async.dart
@@ -52,23 +52,29 @@ class StringListResult {
 @HostApi()
 abstract class SharedPreferencesAsyncApi {
   /// Adds property to shared preferences data set of type `bool`.
+  @async
   void setBool(String key, bool value, SharedPreferencesPigeonOptions options);
 
   /// Adds property to shared preferences data set of type `String`.
+  @async
   void setString(String key, String value, SharedPreferencesPigeonOptions options);
 
   /// Adds property to shared preferences data set of type `int`.
+  @async
   void setInt(String key, int value, SharedPreferencesPigeonOptions options);
 
   /// Adds property to shared preferences data set of type `double`.
+  @async
   void setDouble(String key, double value, SharedPreferencesPigeonOptions options);
 
   /// Adds property to shared preferences data set of type `List<String>`.
+  @async
   void setEncodedStringList(String key, String value, SharedPreferencesPigeonOptions options);
 
   /// Adds property to shared preferences data set of type `List<String>`.
   ///
   /// Deprecated, this is only here for testing purposes.
+  @async
   void setDeprecatedStringList(
     String key,
     List<String> value,
@@ -76,29 +82,38 @@ abstract class SharedPreferencesAsyncApi {
   );
 
   /// Gets individual String value stored with [key], if any.
+  @async
   String? getString(String key, SharedPreferencesPigeonOptions options);
 
   /// Gets individual  void value stored with [key], if any.
+  @async
   bool? getBool(String key, SharedPreferencesPigeonOptions options);
 
   /// Gets individual double value stored with [key], if any.
+  @async
   double? getDouble(String key, SharedPreferencesPigeonOptions options);
 
   /// Gets individual int value stored with [key], if any.
+  @async
   int? getInt(String key, SharedPreferencesPigeonOptions options);
 
   /// Gets individual `List<String>` value stored with [key], if any.
+  @async
   List<String>? getPlatformEncodedStringList(String key, SharedPreferencesPigeonOptions options);
 
   /// Gets the JSON-encoded `List<String>` value stored with [key], if any.
+  @async
   StringListResult? getStringList(String key, SharedPreferencesPigeonOptions options);
 
   /// Removes all properties from shared preferences data set with matching prefix.
+  @async
   void clear(List<String>? allowList, SharedPreferencesPigeonOptions options);
 
   /// Gets all properties from shared preferences data set with matching prefix.
+  @async
   Map<String, Object> getAll(List<String>? allowList, SharedPreferencesPigeonOptions options);
 
   /// Gets all properties from shared preferences data set with matching prefix.
+  @async
   List<String> getKeys(List<String>? allowList, SharedPreferencesPigeonOptions options);
 }